requirements.txt
flask==2.3.3
flask-cors==6.0.1
requests==2.31.0
python-dotenv==1.0.0
google-generativeai